requirements.txt
discord.py==2.3.2
openai==1.12.0